Account for case when property getter or setter has multiple or no matches.
Because property getters and setters can be inherited, the functions may
not be defined in the context that the property is.  This will cause the
lookup to not return any results, and it should not attempt to dereference
the data.  Also, a few places in Qt4 define multiple overloads for a method
that is a property getter or setter (eg. QProgressBar::textDirection).