Add support for passing arbitrary arguments to the clang commandline
MSVC supports the latest C++ standard.  Only as of MSVC 2015 Update 3 was
a /std:c++14 flag added.  There's no way to specify with MSVC 2015 that
you want the code limited to the c++11 standard.  If clang++ is run on
Windows with MSVC 2015 installed, and -std=c++11 is passed, c++14-only
features are included in MSVC headers, and compilation will fail.

Instead of adding commandline arguments for all possible clang arguments,
add an option to control the argv list that is passed to the LibTooling
invocation.